+++ music-service/src/main/java/pl/polsl/controller/MusicAdminController.java
@@ -43,7 +43,11 @@ public class MusicAdminController {
     deleteFileAndMetadata(@RequestParam("id") Long id,
                           @RequestParam("username") String username) {
 
-        Boolean success = musicManagementService.removeFileAndMetadata(id, username);
+        Boolean success = false;
+        if (id == null || username == null) {
+            return ResponseEntity.ok(success);
+        }
+        success = musicManagementService.removeFileAndMetadata(id, username);
         return ResponseEntity.ok(success);
     }
